changed prototype back to int
git-svn-id: https://btanks.svn.sourceforge.net/svnroot/btanks/trunk@5248 5c0bf991-d514-0410-a65f-ce4832e571f3

+++ mrt/random.cpp
@@ -33,7 +33,6 @@
 #	include <stdint.h>
 #endif
 
-//#include "logger.h"
 #include <assert.h>
 
 static uint32_t mrt_rand_seed;
@@ -46,7 +45,7 @@ static uint32_t rand_impl() {
 	return x;
 }
 
-const unsigned mrt::random(const unsigned max) {
+const int mrt::random(const unsigned max) {
 	if (max < 2) 
 		return 0;
 
@@ -62,7 +61,7 @@ const unsigned mrt::random(const unsigned max) {
 	*/
 	x %= max;
 	//LOG_DEBUG(("result: %u of %d", x, max));
-	return x;
+	return (int)x;
 }
 
 void mrt::init_seed() {
